refactor: extract session detail screen into composable components
Break the 527-line monolithic [sessionId].tsx into focused components:
- useSessionDetail hook (data fetching + discriminated union state)
- MainLiftResultsTable (plan-vs-actual with color coding)
- AuxResultsTable (per-exercise tables with timed exercise support)
- SummaryChipsRow (RPE, completion, performance chips)
- TraceButton (prescription sheet with owned state)
- groupAuxSetsByExercise utility

Screen is now a thin 184-line composition layer. Also adds timed
exercise display support in aux tables (fixes #111).
