+++ apps/parakeet/src/modules/session/index.ts
@@ -42,3 +42,9 @@ export { WarmupSection } from './ui/WarmupSection';
 export { WorkoutCard } from './ui/WorkoutCard';
 export { AdjustmentsCard } from './ui/AdjustmentsCard';
 export { SessionContextCard } from './ui/SessionContextCard';
+export { SummaryChipsRow } from './ui/SummaryChipsRow';
+export { MainLiftResultsTable } from './ui/MainLiftResultsTable';
+export { AuxResultsTable } from './ui/AuxResultsTable';
+export { TraceButton } from './ui/TraceButton';
+export { useSessionDetail } from './hooks/useSessionDetail';
+export { groupAuxSetsByExercise } from './utils/groupAuxSetsByExercise';
